Update cell chat actions
